Fix score calculation
Instead of assigning a value of 2 to the first attribute, it should be done in reverse order. Namely, the first attribute should be assigned the largest value (2^n, where n is the number of attributes), and subsequent attributes must be assigned values 2^(n-1), 2^(n-2),... 2.

On the other hand, if the attribute does not increase the pet's score, the increase variable must be decreased regardless to update the next attribute's priority.