Implement non-blocking address to GeoPoint conversion
The used implementation of the Geocoder class only allows synchronous querying, which caused the main thread to block for a long time until results arrived. This in turn caused the PetsTimelineFragment to become unresponsive. In order to solve this, a Thread in charge of performing the query and waiting for the result is now being created and launched every time a Pet's location is obtained.

On the other hand, the synchronous version was also left as is for usage in NearbyPetsFragment, since it makes sense to wait for each Pet's location while displaying them in the map.